refactor(prediction): split PredictionPlots into separate module
Move the PredictionPlots namespace from Prediction.py into a dedicated
Plots.py module, mirroring the adm/ layout (ADMDatamart.py + Plots.py).
Re-export PredictionPlots from prediction/__init__.py so the public API
is unchanged. Also drop unused imports that were carried over from the
original file.

Pure refactor: no behaviour change. All 31 prediction tests pass.
