+++ python/pdstools/prediction/__init__.py
@@ -1,3 +1,4 @@
+from .Plots import PredictionPlots
 from .Prediction import Prediction
 
-__all__ = ["Prediction"]
+__all__ = ["Prediction", "PredictionPlots"]
